@@ -4,14 +4,14 @@ apply plugin: 'io.fabric'
 def gitCommit = 'git rev-parse --short HEAD'.execute([], project.rootDir).text.trim()
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
     defaultConfig {
         applicationId "org.itxtech.daedalus"
         minSdkVersion 15
-        targetSdkVersion 27
-        versionCode 15
-        versionName "1.11.1"
+        targetSdkVersion 28
+        versionCode 19
+        versionName "1.14.0"
 
         buildConfigField "String", "BUILD_TIME", "\"${System.currentTimeMillis()}\""
         buildConfigField "String", "GIT_COMMIT", "\"${gitCommit}\""
@@ -39,19 +39,21 @@ android {
 
 dependencies {
     //Support
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support:support-v4:27.1.1'
-    implementation 'com.android.support:percent:27.1.1'
-    implementation 'com.android.support:cardview-v7:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.1.0-beta01'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.percentlayout:percentlayout:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'com.google.android.material:material:1.1.0-alpha07'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0-alpha06'
     //DNS
-    implementation 'org.pcap4j:pcap4j-core:1.7.3'
-    implementation 'org.pcap4j:pcap4j-packetfactory-static:1.7.3'
-    implementation 'org.minidns:minidns-client:0.3.0'
-    implementation 'com.google.code.gson:gson:2.8.4'
+    implementation 'org.pcap4j:pcap4j-core:1.7.6'
+    implementation 'org.pcap4j:pcap4j-packetfactory-static:1.7.6'
+    implementation 'org.minidns:minidns-client:0.3.4'
+    implementation 'com.google.code.gson:gson:2.8.5'
+    implementation 'com.squareup.okhttp3:okhttp:3.14.1'
     //Analytics
-    implementation 'com.google.firebase:firebase-core:15.0.2'
-    implementation 'com.crashlytics.sdk.android:crashlytics:2.9.2'
+    implementation 'com.google.firebase:firebase-core:16.0.9'
+    implementation 'com.crashlytics.sdk.android:crashlytics:2.10.1'
 }
 
 apply plugin: 'com.google.gms.google-services'
