@@ -3,11 +3,11 @@ package org.itxtech.daedalus.fragment;
 import android.app.AlertDialog;
 import android.os.Bundle;
 import android.preference.EditTextPreference;
-import android.support.design.widget.Snackbar;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
+import com.google.android.material.snackbar.Snackbar;
 import org.itxtech.daedalus.Daedalus;
 import org.itxtech.daedalus.R;
 import org.itxtech.daedalus.activity.ConfigActivity;
@@ -122,11 +122,5 @@ public class DNSServerConfigFragment extends ConfigFragment {
         }
 
         return true;
-    }
-
-    @Override
-    public void onDestroy() {
-        super.onDestroy();
-        Daedalus.configurations.save();
     }
 }
