@@ -2,22 +2,25 @@ package org.itxtech.daedalus.service;
 
 import android.annotation.TargetApi;
 import android.app.Notification;
+import android.app.NotificationChannel;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.net.VpnService;
 import android.os.Build;
 import android.os.ParcelFileDescriptor;
-import android.support.v4.app.NotificationCompat;
 import android.system.OsConstants;
 import android.util.Log;
+
+import androidx.core.app.NotificationCompat;
+
 import org.itxtech.daedalus.Daedalus;
 import org.itxtech.daedalus.R;
 import org.itxtech.daedalus.activity.MainActivity;
 import org.itxtech.daedalus.provider.Provider;
-import org.itxtech.daedalus.provider.TcpProvider;
-import org.itxtech.daedalus.provider.UdpProvider;
+import org.itxtech.daedalus.provider.ProviderPicker;
 import org.itxtech.daedalus.receiver.StatusBarBroadcastReceiver;
 import org.itxtech.daedalus.util.Logger;
 import org.itxtech.daedalus.util.RuleResolver;
@@ -28,6 +31,7 @@ import java.net.Inet6Address;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
 import java.util.HashMap;
+import java.util.Set;
 
 /**
  * Daedalus Project
@@ -47,6 +51,8 @@ public class DaedalusVpnService extends VpnService implements Runnable {
     private static final int NOTIFICATION_ACTIVATED = 0;
 
     private static final String TAG = "DaedalusVpnService";
+    private static final String CHANNEL_ID = "daedalus_channel_1";
+    private static final String CHANNEL_NAME = "daedalus_channel";
 
     public static String primaryServer;
     public static String secondaryServer;
@@ -84,10 +90,21 @@ public class DaedalusVpnService extends VpnService implements Runnable {
 
                         NotificationManager manager = (NotificationManager) this.getSystemService(Context.NOTIFICATION_SERVICE);
 
-                        NotificationCompat.Builder builder = new NotificationCompat.Builder(this);
+                        NotificationCompat.Builder builder;
+                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                            NotificationChannel channel = new NotificationChannel(CHANNEL_ID, CHANNEL_NAME, NotificationManager.IMPORTANCE_LOW);
+                            manager.createNotificationChannel(channel);
+                            builder = new NotificationCompat.Builder(this, CHANNEL_ID);
+                        } else {
+                            builder = new NotificationCompat.Builder(this);
+                        }
 
-                        Intent nIntent = new Intent(this, MainActivity.class);
-                        PendingIntent pIntent = PendingIntent.getActivity(this, 0, nIntent, PendingIntent.FLAG_UPDATE_CURRENT);
+                        Intent deactivateIntent = new Intent(StatusBarBroadcastReceiver.STATUS_BAR_BTN_DEACTIVATE_CLICK_ACTION);
+                        deactivateIntent.setClass(this, StatusBarBroadcastReceiver.class);
+                        Intent settingsIntent = new Intent(StatusBarBroadcastReceiver.STATUS_BAR_BTN_SETTINGS_CLICK_ACTION);
+                        settingsIntent.setClass(this, StatusBarBroadcastReceiver.class);
+                        PendingIntent pIntent = PendingIntent.getActivity(this, 0,
+                                new Intent(this, MainActivity.class), PendingIntent.FLAG_UPDATE_CURRENT);
                         builder.setWhen(0)
                                 .setContentTitle(getResources().getString(R.string.notice_activated))
                                 .setDefaults(NotificationCompat.DEFAULT_LIGHTS)
@@ -99,10 +116,10 @@ public class DaedalusVpnService extends VpnService implements Runnable {
                                 .setContentIntent(pIntent)
                                 .addAction(R.drawable.ic_clear, getResources().getString(R.string.button_text_deactivate),
                                         PendingIntent.getBroadcast(this, 0,
-                                                new Intent(StatusBarBroadcastReceiver.STATUS_BAR_BTN_DEACTIVATE_CLICK_ACTION), 0))
+                                                deactivateIntent, PendingIntent.FLAG_UPDATE_CURRENT))
                                 .addAction(R.drawable.ic_settings, getResources().getString(R.string.action_settings),
                                         PendingIntent.getBroadcast(this, 0,
-                                                new Intent(StatusBarBroadcastReceiver.STATUS_BAR_BTN_SETTINGS_CLICK_ACTION), 0));
+                                                settingsIntent, PendingIntent.FLAG_UPDATE_CURRENT));
 
                         Notification notification = builder.build();
 
@@ -190,9 +207,16 @@ public class DaedalusVpnService extends VpnService implements Runnable {
         stopThread();
     }
 
-    private InetAddress addDnsServer(Builder builder, String format, byte[] ipv6Template, InetAddress address) throws UnknownHostException {
+    private InetAddress addDnsServer(Builder builder, String format, byte[] ipv6Template, String addr) throws UnknownHostException {
         int size = dnsServers.size();
         size++;
+        if (addr.contains("/")) {//https uri
+            String alias = String.format(format, size + 1);
+            dnsServers.put(alias, addr);
+            builder.addRoute(alias, 32);
+            return InetAddress.getByName(alias);
+        }
+        InetAddress address = InetAddress.getByName(addr);
         if (address instanceof Inet6Address && ipv6Template == null) {
             Log.i(TAG, "addDnsServer: Ignoring DNS server " + address);
         } else if (address instanceof Inet4Address) {
@@ -218,7 +242,28 @@ public class DaedalusVpnService extends VpnService implements Runnable {
                     .setConfigureIntent(PendingIntent.getActivity(this, 0,
                             new Intent(this, MainActivity.class).putExtra(MainActivity.LAUNCH_FRAGMENT, MainActivity.FRAGMENT_SETTINGS),
                             PendingIntent.FLAG_ONE_SHOT));
+
+            //Set App Filter
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP && Daedalus.getPrefs().getBoolean("settings_app_filter_switch", false)) {
+                Set<String> apps = Daedalus.getPrefs().getStringSet("filterAppObjects", null);
+                if (apps != null) {
+                    boolean mode = Daedalus.getPrefs().getBoolean("settings_app_filter_mode_switch", false);
+                    for (String app : apps) {
+                        try {
+                            if (mode) {
+                                builder.addDisallowedApplication(app);
+                            } else {
+                                builder.addAllowedApplication(app);
+                            }
+                        } catch (PackageManager.NameNotFoundException e) {
+                            Logger.error("Package Not Found:" + app);
+                        }
+                    }
+                }
+            }
+
             String format = null;
+
             for (String prefix : new String[]{"10.0.0", "192.0.2", "198.51.100", "203.0.113", "192.168.50"}) {
                 try {
                     builder.addAddress(prefix + ".1", 24);
@@ -252,8 +297,8 @@ public class DaedalusVpnService extends VpnService implements Runnable {
             InetAddress aliasSecondary;
             if (advanced) {
                 dnsServers = new HashMap<>();
-                aliasPrimary = addDnsServer(builder, format, ipv6Template, InetAddress.getByName(primaryServer));
-                aliasSecondary = addDnsServer(builder, format, ipv6Template, InetAddress.getByName(secondaryServer));
+                aliasPrimary = addDnsServer(builder, format, ipv6Template, primaryServer);
+                aliasSecondary = addDnsServer(builder, format, ipv6Template, secondaryServer);
             } else {
                 aliasPrimary = InetAddress.getByName(primaryServer);
                 aliasSecondary = InetAddress.getByName(secondaryServer);
@@ -275,11 +320,7 @@ public class DaedalusVpnService extends VpnService implements Runnable {
             Logger.info("Daedalus VPN service is started");
 
             if (advanced) {
-                if (Daedalus.getPrefs().getBoolean("settings_dns_over_tcp", false)) {
-                    provider = new TcpProvider(descriptor, this);
-                } else {
-                    provider = new UdpProvider(descriptor, this);
-                }
+                provider = ProviderPicker.getProvider(descriptor, this);
                 provider.start();
                 provider.process();
             } else {
@@ -287,8 +328,10 @@ public class DaedalusVpnService extends VpnService implements Runnable {
                     Thread.sleep(1000);
                 }
             }
-        } catch (InterruptedException ignored) {
-        } catch (Exception e) {
+        } catch (
+                InterruptedException ignored) {
+        } catch (
+                Exception e) {
             Logger.logException(e);
         } finally {
             Log.d(TAG, "quit");
